Change Buffer.index_of to return f64

+++ src/class/buffer.rs
@@ -90,7 +90,7 @@ extern {
     pub fn includes(this: &Buffer, value: &JsValue, offset: Option<f64>, encoding: Option<&JsString>) -> bool;
 
     #[wasm_bindgen(method, js_name = "indexOf")]
-    pub fn index_of(this: &Buffer, value: &JsValue, offset: Option<f64>, encoding: Option<&JsString>) -> i32;
+    pub fn index_of(this: &Buffer, value: &JsValue, offset: Option<f64>, encoding: Option<&JsString>) -> f64;
 
     #[wasm_bindgen(method)]
     pub fn keys(this: &Buffer) -> Iterator;
